python/requirements.txt
sqlalchemy
tqdm
psycopg2-binary
fastapi
uvicorn
pandas
numpy
ta
scikit-learn
tensorflow
joblib
requests
tabulate
beautifulsoup4
pydantic
selenium
webdriver-manager
transformers